add softer failing to save metadata
Rather than failing if an argument to the metadata is None
(or something along those lines), we should instead just log this
and move on. This metadata isn't really that important, at this
stage (without needing it for restart stuff) and if you happen to
fail here because you don't pass in a value you should, then it
should be on you to look that up in the log and figure it out.

This is only for metadata (mostly associated with the run
parameters) and not the actual results, since those are a bit more
important, and if you can't store something from that it should be
an error or handled specifically (i.e. the new dynesty blobs).