Hotfix/script defaults (#172)
* bring `GCfitter` defaults in line with `main` func

This commit updates a few of the command line argument defaults
to better match those in the core fitter functions, mostly to
be explicitly clear what is being run each time.
This also required setting a couple arguments to values that
I wouldn't ordinarily recommend (dlogz=0.01, for example) because
those arguments are hidden in dictionaries, and can't nicely have
the defaults I would want set in the functions, so instead revert
to the `dynesty` defaults, which are geared for a bit nicer
distributions, I think.
But given that you should probably supply those args when running
anyways, it's not a big deal.

* add softer failing to save metadata

Rather than failing if an argument to the metadata is None
(or something along those lines), we should instead just log this
and move on. This metadata isn't really that important, at this
stage (without needing it for restart stuff) and if you happen to
fail here because you don't pass in a value you should, then it
should be on you to look that up in the log and figure it out.

This is only for metadata (mostly associated with the run
parameters) and not the actual results, since those are a bit more
important, and if you can't store something from that it should be
an error or handled specifically (i.e. the new dynesty blobs).