add support for multiple data colours
This commit allows the passing of a list of `data_color`, instead
of just a single colour, which will allow different datasets to
have their own colours.
The utility of this in many cases, given that the datasets
correspond to a specific mass bin, is debatable, but should be
at least available.