add extra check for vesc/Ndot in fittable models
The somewhat convoluted "defaults" setup for the vesc and Ndot
metadata params (different than age and FeH) require that the
fittable model pass them in as None in order to read them from
observations. If they didn't exist in the observations, however,
this would break. This commit simply moves that initial check
logic into the `FittableModel` init directly.

Really, the default logic should be re-examined, and the defaults
provided in `Model` init only be a worst-case fallback