fix numdens background logic in `plot_all`
Deciding wether or not to plot the number density background
in plot_all, since the numdens plot usually accepts that as a
keyword, was a mess before and often didn't work.
This commit adds a bit more of a robust check, and should
function for any variation of the models.