catch/parse kwargs for combined errorbar & scatter
Various plotting functions use a combination of `errorbar` (to
provided error bars) and `scatter` (to show markers and allow for
seperate colours), however a number of kwargs do not
match between the two. Thus passing a single `**kwargs` would
fail on things like marker sizes.
This commit adds a simple method which parses the passed kwargs
for some of the most common such kwargs and corrects them