allow more (angular) units in `Field`
This commit rewrites a bit of the logic surrounding units in
`Field`. Before the `unit` kwarg was only used to convert some
inputs, but really everything was expected to work in arcmins.
I still expect everything to work in arcmin most of the time, but
it wasn't really necessary to restrict it to that.

The main point was also to allow for the plotting to happen in
different units, which it now mostly supports. Turns out that
I could have just placed the logic into `plot` on it's own,
didn't need to support more units at init, but oh well.