better handle fig creation using given fig with MF
This allows the passing of `show_fields=False` to `plot_mass_func`
when also passing a `fig` created by an old call, which already
does contain a MF field plot.

This is useful for overplotting models without duplicating the
fields. Field axes cannot be added to a figure which was already
created without them