redo numdens background logic & allow subtraction
This commit introduces the `background` argument to the base
plotting functions, allowing us to know subtract a background
value from both the models and data, rather than just plotting
it as a line.
The logic in `plot_number_density` is then redone to support this
and better handle plotting the background instead.