redo & store more individual star type properties
Redoes the logic prieviously defining the different stellar type
bins and storing some of the corresponding properties (mostly
to be used in plotting) such as masses and densities of specific
remnants.
This commit introduces a new named tuple for storing these old
properties and a few new ones for each kind of stellar class.
This includes all remnants together, and all non-BH objects.