fix bug in by-type profile plots for single-models
The profile plots that are based on individual stellar classes
(such as the cumulative mass and density plots) was broken for
single `ModelVisualizer` objects, due to only storing a single
"mass bin" array (because it doesn't make sense to have more than
one mass bin) but still attempting to plot using `star_bin`.

This also now allows you to not apply the mass bin to model
labels, which doesn't make sense in every case anyways.