make `clr_param` support named matplotlib colours
This commit makes the `_add_colours` method within `RunCollection`
more robust to different types of passed colour parameter names.
In the past it was made to support model parameter names, and
somewhat naively supported just passing in arrays of floats, but
did not support the more typical named colours from matplotlib
itself.
In most cases this was not necessary, as those colours should
be passed to the underlying plotting functions (i.e. `color`, `fc`,
etc.), not as `clr_param`, but this does not work for the violin
plotting in matplotlib, so these changes were required to
nicely support that.