add patch for errorbar colours in run relations
when neither `clr_param` or `c` is provided to many of the
run relation plots the colours of the errorbars and the scatter
plot points were not connected. This only really showed up if you
overplotted multiple times, but should now be fixed.
This does seem to sometimes disable the default colour cycle
sometimes, but thats from matplotlib, not my fault