Hotfix/relation legends (#188)
* fix runcollection relation plot legend handlers

Before the handlers passed to the legend from a call to one of the
`relation` functions was handled by the errorbar call, which had
no marker (for various reasons that's handled by a scatter plot
call). This simply moves the label more explicitly to the scatter.
The legend artists will thus have the points, but no error bars.

I think this is a better compromise. Without control over the
legend calls within the function, this is the best that can be
easily done.

* fix colour-cycle bug from #186

Fixes the "bug" introduced in the bugfix by forcing the
errorbar artists to the colour of the scatter points, rather
than vice versa.