correctly coerce string variables to unicode dtype
Since HDF5 cannot handle the UTF-32 fixed-width format (i.e.
numpy dtype "U"), which is the default and recommended way to
store numpy fixed-width strings, we have to coerce any new
`Variable` arrays from byte-string arrays to the "U" dtype.
This is mostly just to get rid of the "bytes", which are rarely
the best choice nowadays