coerce dataframe string columns to "S" bytes dtype
HDF5 is unable to store the typical unicode string objects, and
can only store `bytes` dtype arrays (i.e. the numpy "S" dataype).
This commit simply adds a check, when reading data into a
`ClusterFile` `Dataset`, for dataframe columns which are string
objects, and coerces them to the "S" dtype before trying to write
them to a cluster data file