add some better nested sampler default args (#191)
Reverts a couple of the changes made to the defaults in the
`GCfitter` script by #172 to much better values (i.e. values that
even remotely work) and extends these defaults to the function
as well.
This keeps the sharing of defaults between the two that was the
motivation behind #172, while keeping the original values that
are actually used. As long as how the defaults differ from the
base sampler defaults (in dynesty) are documented, then it's okay.