move MF field sampling to `Observations` init
The Monte Carlo integration of all MF fields used in the MF
likelihoods relies on the (pretty inefficient) sampling of M
points from within each (awkwardly shaped) field slice.
This was done during each likelihood call, but in order to be
fast enough M was set pretty low.
The combination of small M and resampling each time leads to
the value of the MF likelihood, for a given model, to have a
relatively large random spread on each call.

This commit moves the responsibility for the random sampling to
the initial setup of the likelihoods in `Observations`. The
sampling only needing to be done once means that 1) there is no
harm in vastly increasing M, thus increasing the accuracy of the
integral, and 2) essentially, the "random seed" of the integration
is fixed, meaning repeated calls should be identical.