add logic to `Field` to allow pickling
Adds get and set state magic methods which explicitly search for
and remove/re-add any prepped geometries in a `Field` instance
when pickled.
This would add a bit of overhead to re-prep geometries at times,
but should be negligible, and now this class can be passed
through a multiprocessing `map`.