fix number density profiles derived from SB
The number density profiles for three clusters (4372, 4833 & 5927)
don't come from deBoer2019, but from a different compilation and
were given in terms of surface brightness.
The conversion used in the past from SB to number density was
bugged, and had the wrong shape, which hindered fitting. This
simply fixes that and re-adds the number density profiles to those
three clusters.