change legend creation to hide emptyness warnings
Typically when there are no applicable labels to create a
legend a warning saying so is raised. We used to also remove
the blank legend after this.
Unfortunately, this warning seems immune to being caught by
`warnings` somehow, so to avoid it ever happening we'll instead
determine the handles and labels manually, using the same function
as `ax.legend()` uses anyways