update dynesty and switch to normal weight func
Ever since dynesty v2.1.0, likelihood surfaces with plateaus are
supposed to be handled explicitly by their sampler. This should, in
theory, remove the need for the strange custom weight function I
built to (supposedly) get around strangeness with plateaus.
This commit simply shunts that function to the rear in favour of
the default weight function from dynesty.