remove backend "initial batch" logic
As was actually done originally in the stale commit 0345a7, this
commmit removes all custom storage of the on-the-fly results
during the initial batch in favour of just brute storing the
entire results object each iteration.
The idea behind storing the initial batch results in a similar
fashion to the "current batch" results was to avoid messing up the
previous batch's results before they are all combined, but that
clearly doesn't matter in the initial batch as there are no
previous results to mess up, so all that extra logic was not
required, really.

Care should be taken, however, in using any results stored during
this initial generator loop, until it finishes, because some attrs,
like the batch numbers, will be missing until `sample_initial` is
completed.