add tracking of fitting state
This commit adds two new `enum` classes holding flags for the
current state the fitting process is in, which are stored in the backend
intermittently throughout the fitting.

This is just generally nice information to have, but is particularly
useful with the latest changes to the initial batch results storage,
as using any fits stored during that specific loop will probably break
things in `analysis` in very strange ways. This will help alert you
to that (rare) case.