move pool setup to start of fitting functions
Waiting until just before the fitting began to actually initialize
the pool was, I think, leading all the mpi processes to repeat
all of the housekeeping logic at the start of the functions, for
no real reason.
It also became a bug when trying to write the new state metadata
before the pool setup, as all processes tried to open the file at
once.

This commit simply moves the pool setup block to the start of both
fitting functions, to avoid this.