change the HDF file-space management strategy used
The default file space management strategy used is apparently
not free-space persistent, which means that deleted objects were
not having their space re-used (basically). With the new use of
`store_results` throughout the initial batch, this meant a lot of
deletion and in turn could lead to absolutely massive files full of
empty space.

This commit simply touches the HDF file first, with a new fs
strategy, which should alleviate this somewhat.