allow passing arbitrary model args to fitting
This commit modifies the fitting and probability functions to
accept arbitrary keyword arguments to be passed to each model
initialization. This also stores these keywords as metadata and
applies them during any later analysis.

This is mainly meant to handle alterations to clusterBH model
parameters on the fly, but can also be used to change things
related to to e.g. the MF algorithm, or the ODE solver.