overhaul some of run metadata storage
This rethinks how storing metadata containing dictionaries and
lists/arrays is done, in order to support storing nested metadata,
as happen when passing in something like `cbh_kwargs` to the new
`model_kwargs` argument. The largest side-effect to this change is to
do with the stored specified priors, which are actually somewhat
simplified now.
This does unfortunately mean this change is not backwards compatible
when it comes to `_reconstruct_priors`, but that's not used very often.