add catch for died-too-early clusterBH models
Rarely something goes wrong in clusterBH and causes the output
number of timesteps to be very small. This can cause an error
when trying to make a spline out of 2 points.

This shouldn't happen in a normal cluster as the normal timestep
size is 2 Myr and ages are all around 13 Gyr (but some die earlier),
and thus we can probably just safely error here.