split free-BHs into flexible kicks and IFMRs
This takes the recently added optional flexibility in BH conditions
and splits it into two, allowing the flexibility to be isolated to
either the BH natal kicks or the BH IFMR (or both).
No changes to what actually happens when you let these go free,
just allowing you to pick and choose a bit more.