update to latest version of clusterBH
This changes a couple things within the evolved model init to
be able to use the latest version of clusterBH. Namely, this
should work with the updates given in ffronimo/clusterbh#2.

Mostly the only thing that needed to really change was the
computation of the escaping star rate (which is more complicated
now, and which should maybe eventually just come from clusterbh
directly).
The clusterbh parameters have also changed, but the defaults in
`clusterBH` can simply be used as is.