allow for run collection names to come from files
Optionally replaces the typical setting of run names within
`RunCollection` based on the given cluster with the actual stem
of the filename. This allows for easier handling of multiple runs
surrounding the same cluster, without needing to manually change
their names to be unique. When names are not unique, some functionality
is broken, like `filter_runs`.