add further error check for limepy non-convergence
An IndexError can occur in limepy a bit after it has failed to
actually converged to an ODE solution, because it continues on
anyways to computing other quantities. We already check for
convergence, but this IndexError (not ValueError like usual) can
cause the sampler to die, so this extra check was added to catch
when the model errors this way due to non-convergence specifically.

It's a mystery why this has never been a problem in the past.