add more initial BH quantities to model viz
Saves the initial mass, number and fraction of BHs in the
model visualizer classes, in the same way as the present day
quantities were, and with the same compatibility spoofs in the
non-evolved classes as was done with the time-based profiles.

Also finally changes the `BH_mass` and `BH_num` attributes to the
much more obvious `M_BH` and `N_BH` (while hopefully maintaining
backwards compatibility).