fix the computation of `f_BH0` in visualizer
Rather than taking the first value from `f_BH_t`, which incorrectly
double counts the BHs in the calculation of the total mass
(M=M0+Mbh0), this changes `f_BH0` to compute the initial fraction
as it would be done in SSPtools, using M=M0+Mbh0-Ms_lost. There are
reasons clusterBH computes it in this way, but it is confusing when
analyzing the fits. This `f_BH0` is now the "correct" fraction,
and should only be compared with other "correct" fractions