fix small bug in reconstructing prior args
This fixes a small bug when reading in saved string prior args.
The h5py `.astype` method returns a read-only view with conversions
handled by HDF directly, and apparently you cannot simply unpack
these views (or the converter was failing for some reason)
I think this slipped through in the past but seems to error out
on the latest h5py versions.
This fix simply converts the args to numpy first, to avoid this
entirely.