rejig ObservationsVisualizer init
This rethinks the rh and d parameters to the ObservationsVisualizer.
The radius isn't actually needed except for the default signature of
`plot_MF_fields`, so allowing that to be more optional and adding a
better catch in that function makes more sense. The distance is required
but the old implementation worked fine, so I've reverted that, with
the slight change to make `d` a required positional argument, with
the search of 'initials' only happening if explicitly given as None.