update to latest dynesty version
While there are probably some improvements that could be made by
going through the sampler logic again (for the first time in years)
and comparing with dynesty updates, this commit simply updates the
bare minimum required for compatibility for the latest dynesty version,
and makes it a requirement.
While runs made with the old version should still work with
`RunAnalysis`, updating dynesty will be required going forward.